Changes to db , submit quiz

+++ database/migrations/2019_05_02_121347_create_scores_table.php
@@ -19,7 +19,7 @@ class CreateScoresTable extends Migration
             $table->unsignedBigInteger('user_id');
             $table->foreign('chapter_id')->references('id')->on('chapters');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->integer('score');
+            $table->float('score');
             $table->timestamps();
         });
     }
